fix: eliminate possibility of num conversion panics  add: ChecksumBody comments  update: checksum trait method finalize to return result  refactor: parameter order of ChecksumBody::new  add: documents for AwsChunkedBody and AwsChunkedBodyOptions  add: MD5 checksum support to aws-smithy-checksums  update: checksum inlineables

+++ rust-runtime/aws-smithy-types/src/base64.rs
@@ -153,7 +153,7 @@ fn decode_inner(inp: &str) -> Result<Vec<u8>, DecodeError> {
 
 /// Given the length of some data in bytes, return how many bytes it would take to base64 encode
 /// that data.
-pub fn encoded_length(length: usize) -> usize {
+pub fn encoded_length(length: u64) -> u64 {
     (length + 2) / 3 * 4
 }
 
@@ -260,13 +260,13 @@ mod test {
     #[test]
     fn test_base64_encoded_length() {
         let decoded = "Alas, eleventy-one years is far too short a time to live among such excellent and admirable hobbits. I don't know half of you half as well as I should like, and I like less than half of you half as well as you deserve.";
-        let expected_encoded_length = encode(decoded).len();
-        let actual_encoded_length = encoded_length(decoded.len());
+        let expected_encoded_length = encode(decoded).len() as u64;
+        let actual_encoded_length = encoded_length(decoded.len() as u64);
         assert_eq!(expected_encoded_length, actual_encoded_length);
 
         let decoded = "ユニコードとはか？";
-        let expected_encoded_length = encode(decoded).len();
-        let actual_encoded_length = encoded_length(decoded.len());
+        let expected_encoded_length = encode(decoded).len() as u64;
+        let actual_encoded_length = encoded_length(decoded.len() as u64);
         assert_eq!(expected_encoded_length, actual_encoded_length);
     }
 }
